Move auth_request to resources/__init__.py module
